+++ src/components/LexicalFocused/components/Divider/Divider.tsx
@@ -1,0 +1,3 @@
+export const Divider = (): JSX.Element => {
+  return <div className='divider' />;
+};
