+++ src/components/LexicalFocused/components/Divider/index.ts
@@ -1,0 +1,1 @@
+export * from './Divider';
