+++ src/components/LexicalFocused/components/Dropdown/DropDownContext.tsx
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+type DropDownContextType = {
+  registerItem: (ref: React.RefObject<HTMLButtonElement>) => void;
+};
+
+export const DropDownContext = createContext<DropDownContextType | null>(null);
