+++ src/components/LexicalFocused/components/Dropdown/Dropdown.tsx
@@ -1,139 +1,11 @@
 import { isDOMNode } from 'lexical';
-import * as React from 'react';
-import {
-  ReactNode,
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
+import { ChevronDown } from 'lucide-react';
 import { createPortal } from 'react-dom';
+import { ReactNode, useEffect, useRef, useState } from 'react';
 
-type DropDownContextType = {
-  registerItem: (ref: React.RefObject<HTMLButtonElement>) => void;
-};
-
-const DropDownContext = React.createContext<DropDownContextType | null>(null);
+import { DropDownItems } from './DropDownItems';
 
 const dropDownPadding = 4;
-
-export function DropDownItem({
-  children,
-  className,
-  onClick,
-  title,
-}: {
-  children: React.ReactNode;
-  className: string;
-  onClick: (event: React.MouseEvent<HTMLButtonElement>) => void;
-  title?: string;
-}) {
-  const ref = useRef<HTMLButtonElement>(null);
-
-  const dropDownContext = React.useContext(DropDownContext);
-
-  if (dropDownContext === null) {
-    throw new Error('DropDownItem must be used within a DropDown');
-  }
-
-  const { registerItem } = dropDownContext;
-
-  useEffect(() => {
-    if (ref && ref.current) {
-      registerItem(ref);
-    }
-  }, [ref, registerItem]);
-
-  return (
-    <button
-      className={className}
-      onClick={onClick}
-      ref={ref}
-      title={title}
-      type='button'
-    >
-      {children}
-    </button>
-  );
-}
-
-function DropDownItems({
-  children,
-  dropDownRef,
-  onClose,
-}: {
-  children: React.ReactNode;
-  dropDownRef: React.Ref<HTMLDivElement>;
-  onClose: () => void;
-}) {
-  const [items, setItems] = useState<React.RefObject<HTMLButtonElement>[]>();
-  const [highlightedItem, setHighlightedItem] =
-    useState<React.RefObject<HTMLButtonElement>>();
-
-  const registerItem = useCallback(
-    (itemRef: React.RefObject<HTMLButtonElement>) => {
-      setItems(prev => (prev ? [...prev, itemRef] : [itemRef]));
-    },
-    [setItems],
-  );
-
-  const handleKeyDown = (event: React.KeyboardEvent<HTMLDivElement>) => {
-    if (!items) {
-      return;
-    }
-
-    const key = event.key;
-
-    if (['Escape', 'ArrowUp', 'ArrowDown', 'Tab'].includes(key)) {
-      event.preventDefault();
-    }
-
-    if (key === 'Escape' || key === 'Tab') {
-      onClose();
-    } else if (key === 'ArrowUp') {
-      setHighlightedItem(prev => {
-        if (!prev) {
-          return items[0];
-        }
-        const index = items.indexOf(prev) - 1;
-        return items[index === -1 ? items.length - 1 : index];
-      });
-    } else if (key === 'ArrowDown') {
-      setHighlightedItem(prev => {
-        if (!prev) {
-          return items[0];
-        }
-        return items[items.indexOf(prev) + 1];
-      });
-    }
-  };
-
-  const contextValue = useMemo(
-    () => ({
-      registerItem,
-    }),
-    [registerItem],
-  );
-
-  useEffect(() => {
-    if (items && !highlightedItem) {
-      setHighlightedItem(items[0]);
-    }
-
-    if (highlightedItem && highlightedItem.current) {
-      highlightedItem.current.focus();
-    }
-  }, [items, highlightedItem]);
-
-  return (
-    <DropDownContext.Provider value={contextValue}>
-      <div className='dropdown' ref={dropDownRef} onKeyDown={handleKeyDown}>
-        {children}
-      </div>
-    </DropDownContext.Provider>
-  );
-}
 
 export const DropDown = ({
   disabled = false,
