+++ src/components/LexicalFocused/components/Dropdown/index.ts
@@ -1,1 +1,4 @@
 export * from './Dropdown';
+export * from './DropdownItem';
+export * from './DropdownContext';
+export * from './DropdownItems';
