--- src/components/LexicalFocused/plugins/ToolbarPlugin/ToolbarPlugin.css
@@ -1,15 +1,0 @@
-.toolbar {
-  display: flex;
-  border-bottom: 1px solid #eee;
-  background: #fff;
-  padding: 4px;
-  border-top-left-radius: 10px;
-  border-top-right-radius: 10px;
-  vertical-align: middle;
-  overflow: auto;
-  height: 36px;
-  position: sticky;
-  top: 0;
-  z-index: 2;
-  overflow-y: hidden;
-}
